backend/ml/requirements.txt
tensorflow>=2.9.0
numpy>=1.20.0
Pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0 